requirements.txt
h5py==2.10.0
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
tensorboard==2.0.2
tensorboardX==1.9
tensorflow==2.0.0
tensorflow-estimator==2.0.1
torch==1.3.1
torchvision==0.4.2
matplotlib

